build: :arrow_up: Bevy 0.12 & Clippy Linting
This PR aims to add support for Bevy 0.12. While at it this PR also upgrades the other dependencies as well to their latest versions as of the 24th of November. The Code refactoring (idiomization) is a byproduct of my tools which automatically run Clippy. The changes have been tested by simply running the examples.

Bevy version 0.12